pyproject.toml
[project]
name = "aind-import"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
